User correct stage for SSM parameter
